Updated EXE, fixed MSVC warning
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@2451 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/server/CServer_Game.cpp
@@ -997,7 +997,7 @@ void GameServer::gotoLobby(void)
 				cWorms[i].setAFK(AFK_BACK_ONLINE, "");
 				CBytestream bs;
 				bs.writeByte( S2C_AFK );
-				bs.writeByte( i );
+				bs.writeByte( (uchar)i );
 				bs.writeByte( AFK_BACK_ONLINE );
 				bs.writeString( "" );
